changes import to core/resources + renames mispelled filename
