+++ src/components/authorization/dto/beta-features.ts
@@ -1,7 +1,7 @@
 import { Field, ObjectType } from '@nestjs/graphql';
 import { keys as keysOf } from 'ts-transformer-keys';
 import { Calculated, ResourceRelationsShape } from '~/common';
-import { RegisterResource } from '~/core';
+import { RegisterResource } from '~/core/resources';
 import { Granter, ResourceGranter } from '../policy';
 
 // TODO move somewhere else
