--- src/components/authorization/model/resource-map.ts
@@ -1,6 +1,0 @@
-import { AssignableRoles } from '../dto/assignable-roles';
-
-/** @deprecated Use {@link import('~/core').ResourcesHost.getMap} instead */
-export const LegacyResourceMap = {
-  AssignableRoles,
-} as const;
