+++ src/components/field-zone/dto/field-zone.dto.ts
@@ -1,6 +1,6 @@
 import { ObjectType } from '@nestjs/graphql';
 import { keys as keysOf } from 'ts-transformer-keys';
-import { RegisterResource } from '~/core';
+import { RegisterResource } from '~/core/resources';
 import {
   DbUnique,
   ID,
