+++ src/components/partnership/dto/partnership.dto.ts
@@ -1,7 +1,7 @@
 import { Field, ObjectType } from '@nestjs/graphql';
 import { keys as keysOf } from 'ts-transformer-keys';
-import { RegisterResource } from '~/core';
 import { BaseNode } from '~/core/database/results';
+import { RegisterResource } from '~/core/resources';
 import {
   Calculated,
   ID,
