+++ src/components/progress-report/dto/community-stories.dto.ts
@@ -2,7 +2,7 @@ import { keys as keysOf } from 'ts-transformer-keys';
 import { SecuredProps, VariantOf } from '~/common';
 import { RegisterResource } from '~/core';
 import { PromptVariantResponse } from '../../prompts/dto';
-import { ProgressReportHighlight } from './hightlights.dto';
+import { ProgressReportHighlight } from './highlights.dto';
 
 @RegisterResource()
 export class ProgressReportCommunityStory extends PromptVariantResponse<CommunityStoryVariant> {
