+++ src/components/progress-report/dto/progress-report.entity.ts
@@ -7,14 +7,14 @@ import {
   SecuredProperty,
   SecuredProps,
 } from '~/common';
-import { RegisterResource } from '~/core';
 import { BaseNode } from '~/core/database/results';
+import { RegisterResource } from '~/core/resources';
 import { LanguageEngagement } from '../../engagement/dto';
 import { DefinedFile } from '../../file';
 import { IPeriodicReport } from '../../periodic-report/dto/periodic-report.dto';
 import { ReportType } from '../../periodic-report/dto/report-type.enum';
 import { ProgressReportCommunityStory } from './community-stories.dto';
-import { ProgressReportHighlight } from './hightlights.dto';
+import { ProgressReportHighlight } from './highlights.dto';
 import { SecuredProgressReportStatus as SecuredStatus } from './progress-report-status.enum';
 import { ProgressReportTeamNews } from './team-news.dto';
