+++ src/components/progress-report/dto/team-news.dto.ts
@@ -1,8 +1,8 @@
 import { keys as keysOf } from 'ts-transformer-keys';
 import { SecuredProps, VariantOf } from '~/common';
-import { RegisterResource } from '~/core';
+import { RegisterResource } from '~/core/resources';
 import { PromptVariantResponse } from '../../prompts/dto';
-import { ProgressReportHighlight } from './hightlights.dto';
+import { ProgressReportHighlight } from './highlights.dto';
 
 @RegisterResource()
 export class ProgressReportTeamNews extends PromptVariantResponse<TeamNewsVariant> {
