+++ src/components/progress-report/highlights/progress-report-highlights.repository.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { Session } from '~/common';
 import { PromptVariantResponseRepository } from '../../prompts/prompt-variant-response.repository';
 import { ProgressReport } from '../dto';
-import { ProgressReportHighlight as Highlight } from '../dto/hightlights.dto';
+import { ProgressReportHighlight as Highlight } from '../dto/highlights.dto';
 import { oncePerProjectFromProgressReportChild } from '../once-per-project-from-progress-report-child.db-query';
 
 @Injectable()
